feat: add more drivers SattBus, SIXNET, Torque, Triconex, Yaskawa
Add 13 new Priority 5 drivers (#49-61):
- SattBus Ethernet, Scanivalve Ethernet, Simatic TI 505 Ethernet
- SIXNET EtherTRAK, SNMP, Thermo Westronics Ethernet
- Torque Tool Ethernet, Toshiba Ethernet, Toyopuc Ethernet
- Triconex Ethernet, Wago Ethernet, Wonderware InTouch Client
- Yaskawa MP Series Ethernet

+++ src/Atc.Kepware.Configuration/KepwareContracts/Connectivity/OpcUaClient/OpcUaClientDevice.cs
@@ -6,7 +6,7 @@ internal class OpcUaClientDevice : DeviceBase, IOpcUaClientDevice
     public int PublishingInterval { get; set; }
 
     [JsonPropertyName("opcuaclient.DEVICE_SUBSCRIPTION_MAX_NOTIFICATIONS_PER_PUBLISH")]
-    public int MaxNotificationsPerPublish { get; set; }
+    public long MaxNotificationsPerPublish { get; set; }
 
     [JsonPropertyName("opcuaclient.DEVICE_SUBSCRIPTION_UPDATE_MODE")]
     public OpcUaDeviceSubscriptionUpdateModeType UpdateMode { get; set; }
@@ -51,7 +51,7 @@ internal class OpcUaClientDevice : DeviceBase, IOpcUaClientDevice
     public OpcUaDeviceMonitoredItemsDeadBandType DeadBand { get; set; }
 
     [JsonPropertyName("opcuaclient.DEVICE_MONITORED_ITEMS_DEADBAND_VALUE")]
-    public int DeadBandValue { get; set; }
+    public double DeadBandValue { get; set; }
 
     public override string ToString()
         => $"{base.ToString()}, {nameof(PublishingInterval)}: {PublishingInterval}, {nameof(MaxNotificationsPerPublish)}: {MaxNotificationsPerPublish}, {nameof(UpdateMode)}: {UpdateMode}, {nameof(RegisteredReadWrite)}: {RegisteredReadWrite}, {nameof(MaxItemsPerRead)}: {MaxItemsPerRead}, {nameof(MaxItemsPerWrite)}: {MaxItemsPerWrite}, {nameof(ReadTimeout)}: {ReadTimeout}, {nameof(WriteTimeout)}: {WriteTimeout}, {nameof(ReadAfterWrite)}: {ReadAfterWrite}, {nameof(LifetimeCount)}: {LifetimeCount}, {nameof(KeepAliveCount)}: {KeepAliveCount}, {nameof(ConnectionPriority)}: {ConnectionPriority}, {nameof(SampleInterval)}: {SampleInterval}, {nameof(QueueSize)}: {QueueSize}, {nameof(DiscardOldest)}: {DiscardOldest}, {nameof(DeadBand)}: {DeadBand}, {nameof(DeadBandValue)}: {DeadBandValue}";
